refactor: Controller 객체 책임에 맞게 병합
