+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -23,4 +23,9 @@ public class RegisterController {
 
         return new ModelAndView(new JspView("redirect:/index.jsp"));
     }
+
+    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
+    public ModelAndView findRegisterView(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
+        return new ModelAndView(new JspView("/register.jsp"));
+    }
 }
